feat(acp): make the AppKit transcript scroller the only implementation
