--- Alas/Sources/ACP/UI/ACPTopPaginationIndicator.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-enum ACPTopPaginationIndicator: Equatable {
-    case hidden
-    case sentinel
-    case spinner
-}
